Remove duplicate google provider metadata
Previously, `grounding_metadata` and `sources` results from Google API calls
that used the `google_grounding` tool were duplicated in the provider
metadata. Here is an example:

```elixir
iex> {:ok, response} = ReqLLM.generate_text(model, "Search for the latest news", provider_options: [google_grounding: %{enable: true}])
{:ok,
 %ReqLLM.Response{
   error: nil,
   provider_meta: %{
     "google" => %{
       "grounding_metadata" => %{
         ...
       },
       "sources" => [
         ...
       ]
     },
     "grounding_metadata" => %{
       "grounding_metadata" => %{
         ...
       },
       "sources" => [
         ...
       ]
      }
    },
  ...
}
```

It seems like only the metadata withing the "google" key was being used and
tested, and the other addition was unneeded. This PR makes it so that the
metadata is only returned in the "google" key.